Modify tolerance for covariance test.
May need to perform additional research into the
accumulation of floating point rounding error in
this sequence of computations. The 'observations'
are all to the fourth decimal place; the returned
covariance values from the iterative method all
match the 'ground truth' from NumPy to six decimal
places, but not beyond.

Ideally would like to be able to quantify how much
rounding error is expected relative to the input
level of precision.