Add default parameter values to constructor.
Previously the boolean parameters indicated whether to
track variables required to report variance and extrema
did not have default values, as this required each use
to explicitly declare what should and should not be
updated.

Based on use, the most common configuration of tracking
what is needed to report mean and variance, while not
keeping track of extrema, has now been set as the default.